Refactored everything to ES6

+++ src/commands/clear.js
@@ -1,18 +1,34 @@
-module.exports = {
-    name: "clear",
-    description: "clears all the messages in the current channel",
-    permissions: ["MANAGE_MESSAGES"],
-    botpermissions: ["MANAGE_MESSAGES"],
-    dev: true,
+import Command from "../Command.js";
+
+class Clear extends Command {
+    constructor() {
+        super({
+            name: "clear",
+            description: "Clears all the messages in the current channel.",
+            permissions: ["MANAGE_MESSAGES"],
+            botpermissions: ["MANAGE_MESSAGES"],
+            dev: true,
+        });
+    }
+
     async execute(message, args) {
+        let res = super.execute(message, args);
+        if (res) return;
+
         if (message.channel.type == "text") {
-            const n = args[0] || 10;
-            while (true) {                await message.channel.messages.fetch({ limit: n }, true, true);
+            const n = this.args[0] || 10;
+            while (true) {
+                await message.channel.messages.fetch({ limit: n }, true, true);
                 const messages = await message.channel.bulkDelete(n, true);
                 console.log(`deleted ${messages.size} messages!`);
-                if (message.channel.messages.cache.array().length == 0) break;
-                else if (args[0]) break;
+                if (
+                    this.args[0] ||
+                    (await message.channel.messages.cache.array().length) == 0
+                )
+                    break;
             }
         }
-    },
-};
+    }
+}
+
+export default new Clear();
